FastAPI/requirements.txt
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-experimental==0.0.62
langchain-google-genai==1.0.7
langchain-openai==0.1.14
langchainhub==0.1.20
langsmith==0.1.84
python-dotenv==1.0.1